requirements.txt
# TiTiler and FastAPI dependencies for tile server
titiler.application>=0.16
uvicorn[standard]>=0.20.0
fastapi>=0.100.0
python-multipart>=0.0.6

# Rasterio and COG tools for GeoTIFF processing
rasterio>=1.3
rio-cogeo>=5.0

# Configuration management
pyyaml>=6.0